requirements.txt
django
setuptools
pillow
djangorestframework

